code improvements (adding Card & Hand classes)
