+++ exercises/poker/src/example/java/Card.java
@@ -1,0 +1,25 @@
+public class Card {
+	private int rank;
+	private int suit;
+
+	public Card(String card) {
+		rank = parseRank(card);
+		suit = parseSuit(card);
+	}
+
+	public int getRank() {
+		return rank;
+	}
+
+	public int getSuit() {
+		return suit;
+	}
+
+	private static int parseRank(String card) {
+		return "..23456789TJQKA".indexOf(card.charAt(0));
+	}
+
+	private static int parseSuit(String card) {
+		return ".HSDC".indexOf(card.charAt(1));
+	}
+}
