+++ exercises/poker/src/main/java/Poker.java
@@ -1,12 +1,8 @@
 import java.util.List;
-
 import sun.reflect.generics.reflectiveObjects.NotImplementedException;
 
 public class Poker {
-
 	public static List<String> bestHands(List<String> hands) {
-		{
 			throw new NotImplementedException();
-		}
 	}
 }
