+++ exercises/poker/src/test/java/PokerTest.java
@@ -6,16 +6,14 @@ import org.junit.Test;
 public class PokerTest
 {
     @Test
-    public void OneHand()
-    {
+    public void oneHand() {
         final String hand = "4S 5S 7H 8D JC";
         assertEquals(Arrays.asList(hand),Poker.bestHands(Arrays.asList(hand)));
     }
 
     @Test
     @Ignore
-    public void NothingVsOnePair()
-    {
+    public void nothingVsOnePair() {
         final String nothing = "4S 5H 6S 8D JH";
         final String pairOf4 = "2S 4H 6S 4D JH";
         assertEquals(Arrays.asList(pairOf4), Poker.bestHands(Arrays.asList(nothing, pairOf4)));
@@ -23,8 +21,7 @@ public class PokerTest
 
     @Test
     @Ignore
-    public void TwoPairs()
-    {
+    public void twoPairs() {
         final String pairOf2 = "4S 2H 6S 2D JH";
         final String pairOf4 = "2S 4H 6S 4D JH";
         assertEquals(Arrays.asList(pairOf4), Poker.bestHands(Arrays.asList(pairOf2,pairOf4)));
@@ -32,8 +29,7 @@ public class PokerTest
 
     @Test
     @Ignore
-    public void OnePairVsDoublePair()
-    {
+    public void onePairVsDoublePair() {
         final String pairOf8 = "2S 8H 6S 8D JH";
         final String doublePair = "4S 5H 4S 8D 5H";
         assertEquals(Arrays.asList(doublePair), Poker.bestHands(Arrays.asList(pairOf8, doublePair)));
@@ -41,8 +37,7 @@ public class PokerTest
 
     @Test
     @Ignore
-    public void TwoDoublePairs()
-    {
+    public void twoDoublePairs() {
         final String doublePair2And8 = "2S 8H 2S 8D JH";
         final String doublePair4And5 = "4S 5H 4S 8D 5H";
         assertEquals(Arrays.asList(doublePair2And8), Poker.bestHands(Arrays.asList(doublePair2And8, doublePair4And5)));
@@ -50,8 +45,7 @@ public class PokerTest
 
     @Test
     @Ignore
-    public void DoublePairVsThree()
-    {
+    public void doublePairVsThree() {
         final String doublePair2And8 = "2S 8H 2S 8D JH";
         final String threeOf4 = "4S 5H 4S 8D 4H";
         assertEquals(Arrays.asList(threeOf4 ), Poker.bestHands(Arrays.asList(doublePair2And8, threeOf4 )));
@@ -59,8 +53,7 @@ public class PokerTest
 
     @Test
     @Ignore
-    public void TwoThrees()
-    {
+    public void twoThrees() {
         final String threeOf2 = "2S 2H 2S 8D JH";
         final String threeOf1 = "4S AH AS 8D AH";
         assertEquals(Arrays.asList(threeOf1), Poker.bestHands(Arrays.asList(threeOf2, threeOf1)));
@@ -68,8 +61,7 @@ public class PokerTest
 
     @Test
     @Ignore
-    public void ThreeVsStraight()
-    {
+    public void threeVsStraight() {
         final String threeOf4 = "4S 5H 4S 8D 4H";
         final String straight = "3S 4H 2S 6D 5H";
         assertEquals(Arrays.asList(straight), Poker.bestHands(Arrays.asList(threeOf4, straight )));
@@ -77,8 +69,7 @@ public class PokerTest
 
     @Test
     @Ignore
-    public void TwoStraights()
-    {
+    public void twoStraights() {
         final String straightTo8 = "4S 6H 7S 8D 5H";
         final String straightTo9 = "5S 7H 8S 9D 6H";
         assertEquals(Arrays.asList(straightTo9 ), Poker.bestHands(Arrays.asList(straightTo8, straightTo9 )));
@@ -90,8 +81,7 @@ public class PokerTest
 
     @Test
     @Ignore
-    public void StraightVsFlush()
-    {
+    public void straightVsFlush() {
         final String straightTo8 = "4S 6H 7S 8D 5H";
         final String flushTo7 = "2S 4S 5S 6S 7S";
         assertEquals(Arrays.asList(flushTo7 ), Poker.bestHands(Arrays.asList(straightTo8, flushTo7 )));
@@ -99,8 +89,7 @@ public class PokerTest
 
     @Test
     @Ignore
-    public void TwoFlushes()
-    {
+    public void twoFlushes() {
         final String flushTo8 = "3H 6H 7H 8H 5H";
         final String flushTo7 = "2S 4S 5S 6S 7S";
         assertEquals(Arrays.asList(flushTo8), Poker.bestHands(Arrays.asList(flushTo8, flushTo7)));
@@ -108,8 +97,7 @@ public class PokerTest
 
     @Test
     @Ignore
-    public void FlushVsFull()
-    {
+    public void flushVsFull() {
         final String flushTo8 = "3H 6H 7H 8H 5H";
         final String full = "4S 5H 4S 5D 4H";
         assertEquals(Arrays.asList(full ), Poker.bestHands(Arrays.asList(full, flushTo8)));
@@ -117,8 +105,7 @@ public class PokerTest
 
     @Test
     @Ignore
-    public void TwoFulls()
-    {
+    public void twoFulls() {
         final String fullOf4By9 = "4H 4S 4D 9S 9D";
         final String fullOf5By8 = "5H 5S 5D 8S 8D";
         assertEquals(Arrays.asList(fullOf5By8 ), Poker.bestHands(Arrays.asList(fullOf4By9, fullOf5By8)));
@@ -126,8 +113,7 @@ public class PokerTest
 
     @Test
     @Ignore
-    public void FullVsSquare()
-    {
+    public void fullVsSquare() {
         final String full = "4S 5H 4S 5D 4H";
         final String squareOf3 = "3S 3H 2S 3D 3H";
         assertEquals(Arrays.asList(squareOf3 ), Poker.bestHands(Arrays.asList(full, squareOf3 )));
@@ -135,8 +121,7 @@ public class PokerTest
 
     @Test
     @Ignore
-    public void TwoSquares()
-    {
+    public void twoSquares() {
         final String squareOf2 = "2S 2H 2S 8D 2H";
         final String squareOf5 = "4S 5H 5S 5D 5H";
         assertEquals(Arrays.asList(squareOf5), Poker.bestHands(Arrays.asList(squareOf2, squareOf5)));
@@ -144,8 +129,7 @@ public class PokerTest
 
     @Test
     @Ignore
-    public void SquareVsStraightFlush()
-    {
+    public void squareVsStraightFlush() {
         final String squareOf5 = "4S 5H 5S 5D 5H";
         final String straightFlushTo9 = "5S 7S 8S 9S 6S";
         assertEquals(Arrays.asList(straightFlushTo9 ), Poker.bestHands(Arrays.asList(squareOf5, straightFlushTo9 )));
@@ -153,8 +137,7 @@ public class PokerTest
 
     @Test
     @Ignore
-    public void TwoStraightFlushes()
-    {
+    public void twoStraightFlushes() {
         final String straightFlushTo8 = "4H 6H 7H 8H 5H";
         final String straightFlushTo9 = "5S 7S 8S 9S 6S";
         assertEquals(Arrays.asList(straightFlushTo9 ), Poker.bestHands(Arrays.asList(straightFlushTo8, straightFlushTo9)));
@@ -162,8 +145,7 @@ public class PokerTest
 
     @Test
     @Ignore
-    public void ThreeHandWithTie()
-    {
+    public void threeHandWithTie() {
         final String spadeStraightTo9 = "9S 8S 7S 6S 5S";
         final String diamondStraightTo9 = "9D 8D 7D 6D 5D";
         final String threeOf4 = "4D 4S 4H QS KS";
@@ -172,8 +154,7 @@ public class PokerTest
 
     @Test
     @Ignore
-    public void StraightTo5AgainstAPairOfJacks()
-    {
+    public void straightTo5AgainstAPairOfJacks() {
         final String straightTo5 = "2S 4D 5C 3S AS";
         final String twoJacks = "JD 8D 7D JC 5D";
         assertEquals(Arrays.asList(straightTo5), Poker.bestHands(Arrays.asList(straightTo5, twoJacks)));
